LayoutTests/platform/mac/accessibility/model-element-attributes-expected.txt

This tests model elements have basic accessibility support

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Model description: AXDescription: Heart
PASS model.role is 'AXRole: AXGroup'
PASS model.subrole is 'AXSubrole: AXModel'
PASS successfullyParsed is true

TEST COMPLETE
